split worker/mod.rs
